+++ ThunderstoreStuff/CHANGELOG.md
@@ -1,3 +1,11 @@
+# Version 1.5.0
+- Added Setting to make Dropdown Selection persistent
+- Window Position will only get reset if it is out of bounds
+- Changed the background color of the togglebox to improve readability
+- Removed the see-through aspect of light/dark theme (Looked horrendous with the above change)
+- Added some public calls for mod creators (Class: Baum_API)
+- Added basic networking using Photon RPCs (Credits to nickklmao for the PunRPC workaround)
+- Added a call for the Thunderstore API 
 # Version 1.4.5
 - Fixed some broken code
 - Revised the ReadMe/Changelog/Documentation